Move existing metadata-related tests to TestGetSetMetadata
